Default to Exchange configuration
Per (Trac: #1438), Pine now defaults to a configuration suitable
for Exchange.  This includes removing the "INBOX." prefix that Cyrus
required, as well as not disabling GSSAPI.  It also uses
"imap.exchange.mit.edu" as the hostname, not $USER.mail, because
of SSL certificate hostname validation.  We also provide a utility
for the stragglers to switch back to IMAP.